Add support for fallback transitioning.
Fallback transitioning allows a transition to choose not to drive a transition by swapping itself out with another transition instance or a UIKit native transition.

This is commonly used when a transition expects some sort of context to be available, e.g. a context view, and that context is not immediately available. In this case the transition can fall back to a simpler transition or to use the default UIKit transition.

Closes https://github.com/material-motion/transitioning-objc/issues/13