Always add the toView to the container.
On iOS 13, the to view for default transitions has changed from the view controller to a container view.

Prior to this change, this meant that our transitions would not add the "to" view to the view hierarchy, resulting in the transition getting in to a confusing visual state.

After this change, we always add the "to" view to the view hierarchy if it does not already have a parent.